tests: ignore error logged when failed to download the manifest
When Redpanda node is suspended in `RandomNodeOpeartions` test the retry
chain node responsible for manifest download may fail with timeout as
suspension may be longer than the rtc backoff budget. Added the log line
to ignore list to prevent the test to be reported as failed. The
manifest download will be retried.

Signed-off-by: Michał Maślanka <michal@redpanda.com>