tests/rptest: add type stubs for ducktape
Add type stubs for ducktape to enable better type checking of
tests/rptest. These stubs were initially generated using mypy and/or
pyright, and then manually adjusted and updated.

They are incomplete and incorrect in places, but better than nothing.
This does not affect anything at runtime, only type checking in which
may happen in your IDE and in the future in CI.