ssx: make tests fail with early units release
The tests were not strong enough because the function that tested the
mutex state executed within the lambda containing the held units.
Howevever in real world code the function will return a future that may
execute long after the units were released.

This modified test will fail without the fix in the previous commit.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>