kc/configuration: refactored configuration management in kafka client
Previously the `configuraiton` object was a source of configuration for
every part of the Kafka client. It contained properties that were
connection specific, common retry settings but also very specific
consumer and producer settings. Additionally the configuration object
was an instance of `configuration_store` the abstraction that is
designed to hold application configuration.

Refactored Kafka client configuration management to separate connection,
producer and consumer specific setting into separate aggregates. This
way it is going to be easier to reuse parts of the client in other
projects. Additionally exposed api that allow changing the
configuration properties in runtime. Previously it wasn't clear which
properties can be modified during client operation.

Signed-off-by: Michał Maślanka <michal@redpanda.com>