m/record: added very basic defensive check to prevent crashes
Redpada store `model::record_batch` records as opaque bytes. The bytes
are materialized to individual `model::record` lazily when iterating
over the batch records. During the record materialization record header
vector is parsed. The vector size is stored in a variable length encoded
integer. If the data in the record buffer are invalid the size may be
incorrectly deserialized leading to a very large allocation. Added basic
defensive check that compares the requested vector size with the number
of bytes left in the buffer. It the size is greater then the exception
is thrown as we know that header is never smaller than one byte.

Signed-off-by: Michał Maślanka <michal@redpanda.com>