app: proxy class to operate kafka groups from cluster data migrations
Cluster layer cannot depend on kafka layer due to circular dependencies,
so we'll need to use a proxy class with virtual functions.

Seastar's `sharded<Service>` template cannot be declared with base type
but initialized using a constructor of a derived type. It is a bit
annoying, as `ss::shared_ptr<Service>` used inside to store the
instances actually can be initialized with a
`ss::shared_ptr<ServiceSubclass> &&`, but that's not exposed in
`sharded`. Use pimpl when wiring, but pass actual proxy object
references to caller class instances.