`rptest`: lower backoff in `log_compaction_test`
Occasionally, this test would flake due to a non-zero dirty ratio never
being witnessed.

While it seems unlikely that this should be happening, sampling metrics
like `vectorized_storage_log_{closed/dirty}_segment_bytes` which are constantly
changing while compaction is occurring means that there is a non-zero chance
of never witnessing a discrete value above 0.

Reduce the backoff by two orders of magnitude in order to lower the chances of
this flake occurring, and also reduce redundant queries to the `MetricsEndpoint`
that ultimately test the same condition.