kafka: Do not use replicated_partition in produce
The produce request handler uses 'replicated_partition' instead of the
'partition_proxy'. This was fine before because we only had one type of
partition proxy implementation. But this prevents us from introducing
new type of proxy needed for cloud topics.

This commit changes the produce path by replacing
ss::lw_shared_ptr<replicated_partition> with the partition_proxy. When
the 'cloud_topic_partition' will be added the implementation of the
'make_replicated_partition' will be creating the instance of
'cloud_topic_partition' for cloud topics.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>