rpk: Set duplex for m7gd instance types
Introduce the duplex flag and set it for m7gd.

Larger m7gd instance types are not 100% full duplex. Instead they are
somewhere in the middle, i.e.: depending on the workload you might get
100% WIOPS but only 50% RIOPS.

We set it anyway as otherwise we lose out on too much perf and this is
more backwards compatible.

Further this patch clarifies a few comments about iotune behaviour.