`storage`: add `force_compaction` boolean to `self_compact_segment()`
Right now, the self-compaction bitflag for a `segment` is an in-memory
marker of whether a `segment` has already undergone self compaction or not.

In a future commit, this will be removed in place of the persisted
`self_compact_timestamp`, which will be checked instead of the bitflag.

We will need to force self compaction in the case of a `segment` produced via
adjacent segment merging, even if it has a `self_compact_timestamp` set.