c/a/archiver: Start offset_index upload in the background
As long as the segment upload succeeds, the read path can create the offset
index on the fly during segment download, so it's ok to ignore an index upload
failure. This shouldn't affect our ability to advance the collectable index.

With that in mind, this commit pushes the index upload into the background and
adds a small helper function in archiver_fixture to check that the request
was eventually sent.