`storage`: add `maybe_create_contiguous_segment_set()`
Up till now, when we encounter the case in which `segment`s being recovered
have `prev.offsets().get_dirty_offset() >= s.offsets().get_base_offset()`,
we assume the `segment_index` is corrupted (i.e the offsets are incorrect),
and use a `log_replayer` to ensure correctness of the `segment` offsets.

However, after this recovery process, we do nothing to ensure that the
resulting `segment_set` is composed of monotonically increasing offsets,
or rather, that the assumed "incorrect" offsets weren't actually correct to
begin with.

This scenario can easily occur with adjacent merge compaction, in which the
replacement `segment` has been produced from the concatenation operation,
but the `redpanda` broker crashes/goes down before removing all of the
`segment`s that were part of the concatenation. This race condition was
possible before, and has only been made more likely with the recent
enhancements made to adjacent segment compaction to support the concatenation
of more than two `segment`s.

In order to generate a contiguous, non-overlapping `segment_set` from the
`segment`s recovered and those left on disk, add
`maybe_create_contiguous_segment_set()`. This function takes a greedy search
approach in an attempt to prevent any redundant/duplicate data from being
managed in the `log` after `segment_set` recovery is performed. If the
greedy approach fails to create a `segment_set` under these constraints, the
function returns `std::nullopt` and the original deque of `segment`s
will be used instead to build the `log` (i.e the old behavior is preserved).