cloud_metadata: add group offsets upload fallback
Whole cluster recovery relies on a periodic (default: 1h) upload of the
group offsets snapshots created for each consumer group partition. This
is coordinated by the controller leader but the uploads send requests to
the consumer group partition leaders.

Uploading the group offsets for a partition may fail, for example, if
the partition leader is temporarily unavailable. In this case, it is
preferrable to fall back to using the old manifest's snapshots for this
partition instead of uploading a new cluster manifest with the partition
omitted, since having stale data for the groups is preferrable to having
no data at all.