c/p_probe: Reconfigure metrics when scrubbing enabled/disabled
Generally speaking, it doesn't make sense to publish scrubber metrics on
every partition when the scrubber is turned off (a global setting).

Mechanically, this commit adds a config binding for the scrubber config and
a watcher that triggers metric reconfiguration on state change.

Includes a ducktape test to confirm that metrics are indeed published only
when scrubbing is turned ON.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>