cl: Introduce source_topic_syncer
The purpose of this syncer task is to add mirror topics to the cluster
link table.  It will query for a list of topics from the source cluster
and then filter based on filters configured by the user and other
metadata.  If the topic is selected, it is added as a mirror topic to
the cluster link table.

Another task will run to reconcile these updates (to be implemented in a
future PR).

Signed-off-by: Michael Boquard <michael@redpanda.com>