audit: Do not audit ephemeral users
Ephemeral users are used by auditing and schema registry.  There is no
reason to audit the auditing ephemeral user as that does not correspond
to a user derived action.

For schema registry, there are already audit events tracked when the
user access SR via the SR API.  It isn't necessary to audit that the SR
client has performed an action.

Signed-off-by: Michael Boquard <michael@redpanda.com>