cloud_metadata: fix upload_in_term test
The test checked that it was able to patch config, and part of patch
config includes a retry loop if the controller isn't a leader. the retry
loop checked if it were a leader, but it didn't take into account the
possibility of the error indicating that there is current _no_ leader
which is a distinct case. PR adds that to the possible reasons for
continuing the retry.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>