archival: abort mutex wait if scrubber is being stopped
This should fix archival stop hang when an upload is taking longer time.
With this commit, the mutex wait is aborted immediately and we can
proceed stopping the archiver itself which will abort the upload
altogether.

Before this commit, we have to wait for tiered storage upload to
finish. Or for a housekeeping operation to finish.

This is similar to https://github.com/redpanda-data/redpanda/pull/26894