`storage`: pass `abort_source` to `log::start()`
`_log->start()` is called from `consensus::do_start()`. Within `log::start()`
is a call to `offset_translator().sync_with_log()`, which also passes an
`abort_source`. This synchronization function may be a somewhat heavy
operation, as a `log_reader` over the range `(_highest_known_offset, dirty_offset]`
is used to read the `log`.

Previously, the `abort_source` passed to `sync_with_log()` was the
`disk_log_impl::_compaction_as`. However, this isn't very helpful because
this `abort_source` only has `request_abort()` called upon it within
`disk_log_impl::remove()` and `disk_log_impl::close()`- both of which
are called at the _end_ of the typical high-level `partition` removal process
starting from the `partition_manager`.

What this can lead to is the following:

1. `redpanda` is launched. A relatively heavy `partition` with a lagging
   `_highest_known_offset` enters the `sync_with_log()` operation.
2. The user hits ^C to kill `redpanda`.
3. The high level shutdown process is entered from `application.cc`.
4. Within `partition_manager::do_shutdown(partition)`, we first wait
   for the `_raft_manager` to shutdown the `partition->raft()` pointer
   before proceeding to stop the `partition` and then its underlying `log`.
5. In reality, the `raft` pointer hasn't even finished its starting process
   yet, because the `sync_with_log()` process is taking so long!
6. The user waits for `sync_with_log()` to complete, and only then does
   shutdown of the `raft` object finish (as the `_compaction_as` only requests
   abort when the `log` is closed).
7. `log` is closed, `_compaction_as.request_abort()` is called.

Clearly, passing `_compaction_as` to `sync_with_log()` is essentially a no-op,
since it is guaranteed that `sync_with_log()` _has_ to complete before an abort
of `_compaction_as` can be requested.

To make the shutdown process smoother and avoid this sub-optimal behavior, pass
the `consensus::_as` as an argument to `_log->start()` instead. Then, when
an abort is requested of `redpanda` early on, the `sync_with_log()` process
will be correctly terminated early instead of potentially appearing as a long
winded hang.