ct/l1: introduce means to partition objects in metastore
This adds an object_metadata_builder to the metastore that, in future
implementations, can transparently partition data. The idea is to
roughly allow writing code code like the following:

```
map<object_id, object_builder> objects;
auto pending_objects = metastore.object_builder();
for (const auto& tp : tps) {
    // Internally maps domains...
    auto oid = pending_objects.object_id_for(tp);

    auto& object_builder = objects[oid];
    object_builder.add(...read tp...);
    pending_objects.add(oid, ...metadata for tp...);

    if (...object_builder size is sufficient...) {
        pending_objects.finish(oid);
    }
}
// Internally routes to the appropriate domains...
metastore.add_objects(std::move(pending_objects));
```

The interface is expressed to be agnostic to whether the metastore is
actually partitioned, in the hopes that L1 domains can be left as an
implementation detail that isn't exposed to the reconciler, compactors,
etc. An implementaton is added for the simple_metastore that isn't
partitioned at all.

See this thread[1] for discussion that landed us here.

[1] https://redpandadata.slack.com/archives/C08P57TJXRT/p1752858156972319