cloud_topics/batcher: lazily fetch epoch
Eagerly fetching the epoch on startup has some annoying side effects of
temporary errors. So just handle those temporary errors and error the
writes. In practice the error rejecting writes due to no epoch should
only happen if the cluster is unhealthy and the controller is unable
to elect a leader.