application: construct services for L1 metastore
Instantiates the L1 metastore frontend and the RPC service that wraps it
in the Redpanda application, and adds the STM factory that instantiates
the STM for the Raft members of the L1 topic partitions.

This will allow us to implement a metastore client implementation that
interacts with the L1 metastore topic partition leaders (will be in a
later commit).