ct/l1: return corrected next offsets from bad add_objects()
There's general concern that a single partition's bad update can fail a
shared object's commit to the metastore. This case may end up being
quite common, especially if there are leadership changes while
reconciling, as in this case, we end up not being able to replicate LRO,
which may cause subsequent reconciliation attempts to be misaligned.

So, this commit changes the behavior of the metastore to return success
on add_objects() even if some or all of the objects extent ranges are
misaligned with their expected next. Now, a corrected map of partition
to offset is returned, with the expectation that the caller will use the
corrected offsets next time.