direct_consumer/fetcher: unbidden response fix
In fast leadership changes, it is possible for a fetcher to receive a
response for a partition that was not assigned when it snapshots the
partition epochs.

This happens in the following circumstance:
1. reassignment of paritition P from A -> B
2. fetch begins and snapshots epochs
3. reassignment of P from B -> A
4. produce to P
5. A responds with updates for P to fetcher for A

This was causing an illegal memory access because the snapped epochs for
the request would be missing a value for P.

The resolution is to check for not only different epochs but to allow
for absense of epoch in the snapshotted request epochs.

The behavior is functionally the same as an epoch change.