admin-api: add debug API to crash redpanda
Sometimes you just want Redpanda to crash, now. For those times we have
the trigger_crash admin API, which is only available in debug builds of
Redpanda. It can be used to trigger crashes of various types, including
segfaults, aborts, and crashes due to ASAN or UBSAN violations.

We will initially use this in backtrace decoding tests.