rptest: increase SM metrics fidelity
This test relies on Space Management metrics to stop writing data. In
the producer loop we write 512MB at a time and wait for 5s but SM runs
every 30s which in the worst case over-produces couple of GBs. It also
makes the whole test a lot slower than it needs to be.

This should also make the test more reliable by updating
`max_usage_observed` more timely and with higher fidelity information.