iceberg: define some snapshot summary metrics
The Iceberg spec defines some optional snapshot summary metrics to
indicate the changes each snapshot is performing. These summaries are
actually required by certain engines, like Redshift, which will
complain:

```
[2025-08-26 16:51:30] error:  Error parsing table metadata.
[2025-08-26 16:51:30] code:      15003
[2025-08-26 16:51:30] context:   Required field total-records missing.
[2025-08-26 16:51:30] query:     0
[2025-08-26 16:51:30] location:  json_utils.hpp:56
[2025-08-26 16:51:30] process:   padbmaster [pid=1073759619]
```

This commit introduces a few metrics to the snapshot summary that seem
reasonable for an append-only system like Redpanda to update. These will
be used in the append action in a subsequent commit.