ducktape: Add basic CDT net tuner test
While we do have some CDT tuner tests they only test very basic stuff
and mostly effectively just whether the tuner runs without failing.

Given various auto-detection logic in the tuner this is often not good
enough as the tuner might silently do nothing.

In this PR we add a new test for the net tuner which tests the expected
changes in more detail.

For now it's not very generic and only works on a AWS 4-core machine. In
a later patch I will make this more generic and/or verify it's running
on the expected hardware.

The test is not part of any suite. Later we will add a separate suite
and CI pipeline for this as well.