cl/manager: create consumer offsets topic when link needs it
In empty Redpanda cluster `__consumer_offsets` topic is not initially
present. It is created only when the topic is accessed through Kafka
APIs.

In order for migration of consumer groups to work we require topic to be
present as it controls where the consumer group migration task is
running.

Added piece of code that creates the topic whenever a link with consumer
group migrations is created.

Signed-off-by: Michał Maślanka <michal@redpanda.com>