rptest: remove RedpandaServiceBase
This is a vestigial base class: it is not instantiated directly and has
only one subclass, RedpandaService, which should be used directly.

The "true" base class of RedpandaService and RedpandaServiceCloud is
RedpandaServiceABC.

This change merges all static attributes and methods of RS and RSB and
combines their constructor. Some type fixes are made to redpanda.py
as well.