rptest: local_venv.sh tries to detect changes
tests/local_venv.sh is a way to set up a local venv with ducktape
dependencies, so that type information is available to your editor (
among other things). Make it smart about changes to things that might
change the venv, by hashing its inputs and re-running when they have
changed.