ct/l1: support combining objects
Add an API to support combining L1 objects. This can be useful in places
like reconcilation or compaction that might be working on multiple
partitions at once and they want to combine them together after some
parallel processing. This is possible because the container is
essentially a list of list of batches, the only pointers into the data
is in the footer, so that's the only place that needs to be updated.
Technically if we just knew the footer position we could not require it
to be passed in, but the current API assumes you are combining the
direct output of the L1 object builder.

Add a test that shows you get the same output if you build seperately
and combine or if you build the aggregate output using the reader APIs.
This is purely an optimization where we don't have to deserialize the
partition data, but can just copy the bytes around directly.