`config`: don't throw on unknown property in `config_store::read_yaml()`
`read_yaml()` is called from a number of places in the tree, but the behavior
I am presently most concerned with altering is what happens when `redpanda`
is upgraded or is bootstrapped for the first time.

In this case, reading an unknown cluster property from a configuration file
would previously throw an exception and result in an early abort from `application`.

Effectively, `redpanda` could be stuck in a crash loop until the configuration
file was amended to remove the unknown property.

For context, this could happen when a version of `redpanda` with a cluster
configuration present (for example, due to backporting) is upgraded to a version
of `redpanda` without that cluster configuration present.

This effectively means we cannot safely backport changes with cluster
configurations, even if we'd _really_ like to (with good reasons) [1].

Change this code path to no longer throw an exception and abort start up,
but rather log a `WARN` level message about the unknown property.

This means we can now safely backport changes with added cluster configs
to versions of `redpanda >= 25.3.1`.

While the early abort may save some users who have e.g. mistyped a configuration
property in their configs, this sort of behavior is too pessimistic and
causes great headaches for us when we find ourselves in the above situation
(important fix with cluster config is added and we'd really like to backport
it but cannot).

It also more closely resembles the non-bootstrapping code path, which doesn't
early abort on unknown property, but simply logs at `info` level [2].

[1]: https://redpandadata.slack.com/archives/C07HD9U0EHL/p1740619422891279?thread_ts=1740614635.834399&cid=C07HD9U0EHL
[2]: https://github.com/redpanda-data/redpanda/blob/9f2fddafd1f7d1aa22705d1d3ea3dcb0687affce/src/v/cluster/config_manager.cc#L332-L334