ct: cancel pending requests on write_pipeline shutdown
In a recent run of RNOT we observed many instances of

    ducktape.errors.TimeoutError: Redpanda node docker-rp-14 failed to stop in 30 seconds

The logs for rp-14, we see that the last service that was attempted to
shutdown was the L0 write pipeline.

    INFO  2025-09-19 07:05:19,419 [shard 0:main]
    main/cloud_topics::data_plane - sharded_service_container.h:94 -
    Stopping cloud_topics::l0::write_pipeline<seastar::lowres_clock>, ..next
    to shutdown is cloud_topics::l0::cluster_services

The only thing that write_pipeline::stop blocks on is its gate closing.
And the only place we see the gate being held is in write_and_debounce
where requests are put onto a pending list and then awaited for the
request to finish.

Things like the batcher process requests from the pending list. In the
shutdown sequenece we see that the batcher is stopped before the write
pipeline. Maybe that order is ideal or not, but in general when the
write pipeline stops it appears to be waiting on never-to-be-resolved
requests.

This PR resolves this by resolving any pending requests with an
appropriate error message when the write pipeline is stopped.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>