ct: allow pending objects to be removed from builder
A pattern is used in the reconciler by which the metastore object
builder initializes many objects within the same builder, reconciles
partitions into these objects, and then finishes/commits them.

When only a single object is processed at a time then cleanup is simple
because the entire builder can be thrown away. However, we don't want to
throw away the entire builder because a subset of partitions had an
error during reconciliation.

In this case we need to remove objects in the pending state, as down
stream processing of the builder expects that the builder is in a state
in which all of its objects are "finished".

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>