tuner: Support dedicated net tuner mode in rpk:start
This adds the second half of the dedicated net tuner mode where `rpk
start` will automatically start redpanda with the right cpuset such that
it's not running on the interrupt cores.

This is an early commit which supports the basic functionality. We will
later need to add a lot more checks to not enable this when not tuned
correctly.