tuner: Add flag to enable/disable dedicated mode
Disable the automatic dedicated mode by default.

This will allow us to merge the new mode even while we don't have all
the checks in place yet that are needed for a full default roll out.

Further it will simplify the rollout of this in our cloud where we run
the tuner but also do some manual changes/hacks after that. With the
extra flag less coordination is required.

We can change the default of this flag once the checks are in place but
likely only after 25.3 has branched to get some more data/experience as
we are already close to code completition.