ct: Simplify batcher
Previously, the batcher was responsible for aggregating and uploading
the data but also for scheduling the uploads. It accumulated write
requests over time and triggered the upload.

This commit changes this behavior. The batcher is now responsible only
for uploading and aggregating the data. When the write request is
propagated to the batcher it triggers upload without waiting. It may
bundle multiple write requests if there are many write requests in the
pipeline.

The ide here is that the 'write_request_scheduler' is responsible for
scheduling now so the batcher should just react to signals from the
scheduler. The scheduler accumulates enough data for the upload and then
it pushes write requests to the batcher atomically so the batcher "sees"
many requests after it is notified.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>