tuner: Add RX queue count tuner
Adds a tuner to lower the RX queue count in case we have less interrupt
cores than available queues.

The scylla DB perf_tune already does this.

perf_tune's logic is a bit different in that it just tries to lower both
of RX and Combined queues and is happy as long as one succeeds.

We just lower either (or both) depending on which max is greater than
zero.