test.bzl: default seeding policy to fixed
Set the default RNG seeding policy to fixed for tests. This means that
tests will still used fixed seeding as desired, by default, but that
redpanda application will use random.

Test runs can also opt-in to random, which we will do for runs that are
not tied to a specific change.