cl/tests: added cluster linking tests executing random operations
Added cluster linking test that is inspired by the
`RandomNodeOperationsTest`.

The test uses node operations executor, admin fuzzer and failure
injector to introduce changes in either source or target cluster.

The test includes different topics, each topic represents a different
type of workload:
- cloud storage enabled topic
- compacted topic
- topic with transactions

Each topic has a producer in the source cluster and a consumer in the
target cluster. This way a verifier is able to detect inconsistency
issues and at the same time it enables validating consumer group
migrations.

Signed-off-by: Michał Maślanka <michal@redpanda.com>