cl/topic_syncer: always set remote allow gaps property
When using shadow linking target cluster has no control and knowledge
about the source topic content. It may contain gaps that were either
created by the compaction working in the past or aborted transactions.

In order to allow shadow linking work with tiered storage enable topics
and replicate topics with gaps in the offset range we must allow remote
gaps in every shadow topic that is created in the target cluster.

Added simple mechanism to set overrides for topic properties in topic
syncer.

Signed-off-by: Michał Maślanka <michal@redpanda.com>