cs/archiver_service: relax metadata consistency check when gaps allowed
When topic is configured to tolerate gaps the segment metadata
consistency check should also allow gaps to exists at the upload chunk
boundaries. This commit relaxes ntp archiver metadata check when gaps
are allowed to exists in the remote data.

With the relaxed check the logic doesn't require start and end offset of
stats to be exactly equal to the values coming from upload metadata but
the range defined by the stats offsets must be enclosed within the range
defined by segment metadata.

Rationale:

The consistency check compares the offsets collected when selecting an
upload candidate with the offsets obtained during the full read of the
range selected to upload. Since the batches are actually read when
building remote segment index the statistics collected during the build
are treated as a source of truth for the check.

When gaps exists in the log it may happen that the segment index builder
will skip some batches when building the index. When the gap exists
either at the begging or the end of upload range the metadata
consistency check will fail.

The relaxed check applied when topic property allows remote data gaps to be
present is changed in a way to make the gap at the beginning and the end
of read range not to fail the consistency check.

Gap in the range will not influence the upload result but it will not be
reflected in the stats collected while building remote segment index as
the reader will not 'see' a batch with gap offset.

This way the base offset of stats coming from index creation will always
be greater than the segment metadata base offset if a gap is at the
begging of the range. The analogical rule applies to the last offset of
segment stats.

Signed-off-by: Michał Maślanka <michal@redpanda.com>