ct/l1: add topic_purger
Adds a topic purger in the metastore subsystem. This will check the
topic table for tombstoned cloud topics, kick off topic removals in the
metastore, and then remove the tombstones in the controller.

This commit only adds the component that does the operation in a
one-shot action. A later commit will add a loop to run it periodically.