`storage`: make `adjacent_merge_compact` pass by value
This could lead to a seg fault due to iterator invalidation by concurrent
mutation of the underlying `_segs` container while it was being iterated
over within `adjacent_merge_compact`.

Change the function signature to take `segment_set` by value, forcing
the user to perform a copy of the segments container to avoid this
issue.