`ct/l1`: use linearized queue for handling of backgrounded `_work_fut` future
It was possible before for a pause/resume request to a worker to race
due to the use of `std::exchange` leaving `_work_fut` in a disengaged
state, but perhaps not a _resolved_ state.

To make the process of concurrent pausing/resuming easier to reason about,
linearize the process using an `ssx::work_queue`.