k/server: closed connection tracking fixups
Fix use after move across multiple ListKafkaConnections endpoint calls
by making the data type wrap const objects.

And treat the connection as aborting until `to_closed_proto()` is
called, because otherwise it is possible to observe a race where the
abort source is closed (and therefore the state is reported as closed),
but stop() has not completed yet and therefore the `close_time` is
unset. Avoid this by reporting the connection as aborting in this state.
This was highlighted by flakiness in the existing test checking for the
closed connection.