c/leaders_table: protect partition leaders table with mutex
Leaders table mutations and async iteration may race. In this case the
iteration triggered an assertion as the iterator was dereferenced after
the mutation happened. In theory partition leaders table has a mechanism
to throw when such situation happen but when using `async_for_each` the
iterator is dereferenced before the check is executed laeding to an
assertion.

Added a mutex protecting state while iterating over leaders and
underlying container mutations. The leaders table is not on the hot path
and it is perfectly fine to make mutations and iterations mutually
exclusive.

Signed-off-by: Michał Maślanka <michal@redpanda.com>