ct: Handle shutdown and timout in L0 reader
Previously, the reader was throwing exceptions in case of any error. It
was impossible to discriminate the error because every error was a
download_error. Now the reader handles shutdownn and timeout errors
differently. It returns empty result.

This is needed to handle timeouts properly. The timeout will trigger
timeout error in the fetch handler which will be propagated to the
reader. We should return empty or accumulated previously result in this
case.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>