dt: Handle out-of-order /proc/interrupts in GCP tests
We saw GCP CDT net tuner tests flake because of a peculiar
/proc/interrupts:

```
28:  ...     PCI-MSIX-0000:00:05.0   8-edge      virtio1-output.3
29:  ...     PCI-MSIX-0000:00:05.0   0-edge      virtio1-config
30:  ...     PCI-MSIX-0000:00:05.0   1-edge      virtio1-input.0
31:  ...     PCI-MSIX-0000:00:05.0   2-edge      virtio1-output.0
32:  ...     PCI-MSIX-0000:00:05.0   3-edge      virtio1-input.1
33:  ...     PCI-MSIX-0000:00:05.0   4-edge      virtio1-output.1
34:  ...     PCI-MSIX-0000:00:05.0   5-edge      virtio1-input.2
35:  ...     PCI-MSIX-0000:00:05.0   6-edge      virtio1-output.2
36:  ...     PCI-MSIX-0000:00:05.0   7-edge      virtio1-input.3
```

Note how the the last output queue somehow has a lower IRQ id than the
others.  The expectations rely on this list to be ordered. Hence, change
the test such that it orders the list based on the matcher. Also adjust
the GCP regex such that we get the required ordering.

The tuner already does this right based on the queue index functions
which also use regexes.