type-check.py: fix --update
Recently changes to move the type-checker out of tests and into tools
broke the --update function when using via docker becuase we no longer
mount the type-checker directory into the container. Fix this by
mounting them into the container and using an explicit --entrypoint in
the invocation (and removing that from the dockerfile) as the parth is
now a user path and so is variable.

We also need to update the github action: we just switch to using the
type-check.sh to DRY it (which then picks up these fixes automatically).