admin: rate limit ListKafkaConnections
Rate limit the endpoint to 1 external cluster-wide request per broker.
Note that each cluster-wide request turns into 1 sub-request per broker,
so the rate limit allows up to num_brokers in flight requests on each
broker at the same time.

This rate limiting is to limit the CPU impact of the endpoint, a
separate memory semaphore is introduced in the next commit.